changed spec and got regeneration to work

+++ schooladminimpl.go
@@ -131,7 +131,7 @@ func createTeacher(db *bolt.DB, newUser UserInfo) (err error) {
 }
 
 func createStudent(db *bolt.DB, newUser UserInfo, pathId PathId) (err error) {
-	newUser.Income = int32(math.Floor(rand.Float64()*(335-104) + 104))
+	newUser.Income = float32(math.Floor(rand.Float64()*(335-104) + 104))
 	newUser.CareerTransition = false
 	err = db.Update(func(tx *bolt.Tx) error {
 		err = AddUserTx(tx, newUser)
